[FIX] HTML mode: QUnit.config.noglobals with coverage enabled
Code instrumented by istanbul introduces global variables
(cov_* for each file and a global __coverage__ variable).
This causes a failure when the "noglobals" QUnit config is active and
some file is loaded during a test as this pollutes the global namespace.

Hooking into the "after" function allows to add those variables to the
list of known globals so that they won't be detected as new globals.