[FIX] Ensure defined urlParameters order
Some testsuites may define url parameters for a test and expect the
order of parameters to be unchanged.

It is not fully clear for which cases the sorting has been implemented
initially but it has been decided to rather check for those issues
instead of always sort to have the same order in every browser.

BCP: 2370044253